src/main/java/com/device/qa/controller/UserController.java
package com.device.qa.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.device.qa.model.User;
import com.device.qa.services.UserService;

@RestController
@RequestMapping("/api/user")
public class UserController {

    @Autowired
    private UserService userService;

    @PostMapping
    public ResponseEntity<User> register(@RequestBody User user) {
        return ResponseEntity.ok(userService.save(user));
    }
    
    @GetMapping(path = "/{user-id}")
    public ResponseEntity<User> findById(@PathVariable(name = "user-id", required = true) Long userId) {
        return ResponseEntity.ok(userService.findById(userId));
    }
   

}
